mononoke/repo_client: refactor MononokeRepo (the hg server kind) construction
Summary:
Right now, instantiating a new MononokeRepo is entangled with instantiating a
repo handler that uses it, and doing cache warmup, and starting hooks.

I'd like to be able to create a new repo client without all the surrounding
functionality to build a new traffic replay mechanism that calls directly into
a RepoClient. To do this, I need to be able to construct just a MononokeRepo.

To do so, let's refactor the creation of MononokeRepo into a
MononokeRepoBuilder. It's a builder instead of being just a function for 2
reasons:

- When setting up a repo, we want to start cache warmup as early as possible,
  which means as soon as we have a BlobRepo.
- When setting up a repo, we can get a BlobRepo without a HookManager, but then
  we neeed a HookManager to proceed further.

To do this, the MononokeRepoBuilder can be prepared at which point you can get
a BlobRepo, and then you finalize it

As part of this, I  refactored a few further things:

- Rather than have a ad-hoc function to open a DB from config in repo_listener,
  that's now part of sql_ext.
- We now dispatch cache warmup before we have hooks ready, by putting it on its
  own task. This future is already `'static + Send` so that's easy.
- This uses async / await now.

Reviewed By: StanislavGlebik

Differential Revision: D19553400

fbshipit-source-id: 019760b9143b07b38fc6fed1537847f8d2ec4aec