mononoke: do not spawn many gettreepack streams
Summary:
We ran into memory issue when processing gettreepack request with lots of
mfnodes.

After reproducing it locally I noticed that the problem is in select_all()
method. It spawns a separate stream for each mfnode, and it causes a huge
memory overuse. That makes sense - each stream internally uses
bounded_traversal() which put the limit on the number of simultaneously
processed futures and hence limits the memory. By doing select_all() on these streams
we effectively multiply the total memory used by the number of mfnodes in the
request.

To fix it I suggest to remove select_all() and process at most one stream at a
time. I ran experiments with a few options - one stream at a time,
select_all() on 4 streams at a time and select_all() on all streams. One stream
at a time proved to be the most efficient one. Single mfnode should also be the
most common case, though I'll run a canary to make sure we don't regress
performance.

Reviewed By: krallin

Differential Revision: D19577140

fbshipit-source-id: 8b42984bbec8ac918eab8ffb2faf0666f1271101