mononoke/pushrebase: split PushrebaseParams
Summary:
Right now, PushrebaseParams contains a set of flags that the pushrebase crate
considers, and some flags that unbundle considers. This is a bit messy. Let's
split them into those the pushrebase crate considers (PushrebaseFlags), and
those unbundle considers.

This will make adding Globalrevs in there a little nicer, since pushrebase can
consider Globalrevs by way of a hook, so it would be confusing to receive
PushrebaseParams that mention Globalrevs but not look at it.

Reviewed By: HarveyHunt

Differential Revision: D19516669

fbshipit-source-id: 2f2f9023ad65611f7d71da968ab48d15bd4071dd