mononoke/pushrebase: introduce Pushrebase hooks
Summary:
This introduces the notion of Pushrebase hooks. Those are hooks that can run at
3 stages during pushrebase:

- When we're about to kick off an iteration
- When we're producing new commits
- When we're updating bookmarks

There are going to be 2 users of this for now:

- Commit rewriting, which wants to insert things in its mapping (so just mostly
  the 3rd stage here).
- Globalrevs (implemented later in this stack), which wants to look up the max
  globalrev (1st stage), populate commit extras (2nd stage), and insert them in
  the same transaction as bookmarks (3rd stage).

This is also convenient to e.g. simulate races in tests (see later in this
stack).

The way this is implemented, hooks effectively return a new hook every time
they go through the stages above, which makes this interface harder to misuse
(e.g. you can't accidentally use the wrong max globalrev when assigning it to
commits unless you're really going out of your way to fail):

- You start with a PushrebaseHook
- When an iteration of pushrebase starts, your PushrebaseHook must provide a
  PushrebaseCommitHook.
- When the commits are created, your PushrebaseCommitHook can modify them.
- When the commits have all been created your PushrebaseCommitHook must provide
  a PushrebaseTransactionHook.
- When a bookmark write is about to happen, your PushrebaseTransactionHook must
  populate the transaction.

The API treats hooks as `Box<dyn ...>` so that you can have as many types as
you want.

There are a lot of things in pushrebase that could probably be rewritten to use
this framework. For example, tracking bundle replay data. I haven't done so
just yet, since the API is a little verbose at this point. If we'd like to port
more things to Pushrebase hooks, we can easily add helpers to make it less
verbose.

NOTE: There is a bit of duplication happening in the bookmark move functions.
I've fixed this later in this stack.

Reviewed By: StanislavGlebik

Differential Revision: D19516661

fbshipit-source-id: 5aff934712767ff5691a7269dfd7aeb869af37ca