mononoke/bonsai_globalrev_mapping: expose root cause of add_many failures
Summary:
This updates add_many to return an Error that explains why the add failed, and
lets callers disambiguate between a server error and a conflict.

I was hoping to use the unique index we have for this, but unfortunately that's
not really feasible right now since we don't expose a matchable error from our
DB when we hit a conflict. So, for now, let's use the same pattern we use
in bookmarks to detect conflicts, which is to count affected rows on insert.

Reviewed By: HarveyHunt

Differential Revision: D19516662

fbshipit-source-id: 4a94eebec24724d05190bba0f9b03e1fbb4850cb