mononoke/fastreplay: switch to Cow<'a, str> for borrowed data
Summary:
We're deserializing JSON, and some data will be borrowed (if it can be accessed
straight from the JSON), but encoded strings (e.g. in `"foo": "\"quoted\""`)
won't. Using `str` works for the former, but fails for the latter. Using
`String` works for both, but it allocates a String even when we don't need one.
Using `Cow` does the right thing in both cases.

For reference: https://github.com/serde-rs/serde/issues/914

Reviewed By: StanislavGlebik

Differential Revision: D19577791

fbshipit-source-id: a4eee7a9c1d771a2b0760daeaa6bf8dd0c6b8fbb