mononoke/fastreplay: add integration tests
Summary:
This adds integration tests for fastreplay (adapted from our existing traffic
replay test, which Stas had added a little while ago).

This works by adding support for logging to files in wireproto logging, which
we already have in a bunch of places for Scuba logging and has turned out to be
very convenient.

Since fastreplay operates directly on the JSON we generated from wireproto
logging, we can just pipe it back in, and it all works.

Reviewed By: StanislavGlebik

Differential Revision: D19577792

fbshipit-source-id: 99e6f12f5c06e7a4a69df08c0843c699b639ce54