mononoke: derive filenodes when returning getbundle response
Summary:
We want to derive filenodes so that by the time a client sends gettreepack/getfiles request the filenodes are already derived.

Because of that deriving them at the time we return getbundle response seems like an appropriate place - until new commit is returned clients don't know the hash of the root manifest hence they can't request filenodes that might not be generated yet.

The important change in behaviour is that now we always calculate phases, while previously we were doing it only if client requested them. I'll run a canary to see the impact of this change.

Reviewed By: krallin

Differential Revision: D19519242

fbshipit-source-id: c513a66141ee895c33a5e6c7cb6048a9fff760c3