mononoke: add cachelib support for phases
Summary:
This support is added with the help of GetOrFillMultipleFromCacheLayers struct
(the same we use for bonsai_hg_mapping and changesets).
It checks two caching layers before going to db, and fills them up
appropriately.

I've also added a method to just fill the caches which is useful when we write
new phases to the db

Reviewed By: farnz

Differential Revision: D19580655

fbshipit-source-id: 544fd1532727da209474fe73b4eaebf4a54fdaa7