mononoke: put phases in BlobRepo
Summary:
Previously Phases object wasn't stored in BlobRepo, and the reason was because Phases depended on BlobRepo.

That wasn't great because we needed to have a separate initialization for phases in all the places that needed them (e.g. repo_client, blobimport etc).

This diff reverses the dependency (i.e. now BlobRepo depends on Phases, not the other way around).

To achieve it a few adidtional objects were added:
1) HeadsFetcher - just a typedef for a function that returns all public heads in the repo
1) SqlPhasesStore - raw interface to db
2) SqlPhases and CachingPhases - objects that implement Phases interface.
3) SqlPhasesFactory - and object that returns Arc<dyn Phases> (i.e. either SqlPhases or CachingPhases) given changeset fetcher and heads fetcher

Reviewed By: krallin

Differential Revision: D19569688

fbshipit-source-id: c260a573d48caba89126287dfd189cf2e117ceea