py3: update ui.setconfig() to work in Python 3
Summary:
The `ui.setconfig()` code attempts to escape lists, and checks for them by
looking if the config value has an `__iter__` attribute.  However, in Python 3
both `str` and `bytes` types have an `__iter__` method, so the setconfig
code incorrectly treated them as a list.

This re-orders the code so that `str` types are explicitly handled before the
check for iterable types.

Reviewed By: xavierd

Differential Revision: D19596740

fbshipit-source-id: 3c52d3e8d5b68fd2d561651bd1858889cdbe0ccb