py3: pycompat: make fsencode/fsdecode do nothing
Summary:
We want to assume almost all paths are unicode and drop support for non-unicode
paths. `os.fsencode` returns bytes that breaks the assumption.

Change fsencode and fsdecode to do nothing so path types remain unicode.

Most use-cases are related to `__file__`, which is already unicode on Python 3.
`posix._checklink` uses it. But it seems making it no-op is okay.

Reviewed By: farnz

Differential Revision: D19613693

fbshipit-source-id: 587dbeef68b23beeaf6bc8861c71ea27b3c512d7