hide duplicate hg error messages in eden doctor output
Summary:
If the hg data in a checkout is corrupt running `hg` commands may cause
Mercurial to crash and print backtrace information.  Hide this error output
when performing our `check_filesystems` checks.

The `check_hg` logic will already detect and handle corrupt repositories.  If
the hg error messages are displayed directly to stderr when running the
`check_filesystems` checks earlier this just makes the output more confusing
to the user.

Reviewed By: genevievehelsel

Differential Revision: D19629960

fbshipit-source-id: 774287f37b216216be3eda2247307416d7ae4034