mononoke/fastreplay: don't load remote args in the main thread
Summary:
Right now, the main thread does everything up to the point of dispatching a
request to a RepoClient. This is nice for performance, because it avoids
having to `Arc<...>` any of fastreplay's datastructures, but it's not great for
scalability, since it means we cannot be preparing more than one request for
dispatch in parallel.

Spawning earlier should avoid that problem. This comes at the cost of `Arc`-ing
some things, but in the grand scheme of things that does not matter.

Reviewed By: HarveyHunt

Differential Revision: D19639494

fbshipit-source-id: 1cd40005f0382593c00066751a20ca76ff345d56