mononoke: create cmdlib_x_repo library
Summary:
For now it has only a single function to create a commit syncer. I could have put it in cmdlib, however I
don't think it makes to add a dependency on cross_repo_sync for all libraries
that depend on cmdlib. Hence a separate binary.

This function will be used in the next diff

Reviewed By: krallin

Differential Revision: D19623842

fbshipit-source-id: 630733e0077309b22221bcef0fe758fa24b507f9