mononoke: Allow control over cachelib's HashTable bucket size
Summary:
Cachelib uses a HashTable to map keys to cache objects. The size of
this HashTable can be configured to suit the number of keys that are stored in
the cache. This is a trade off between memory usage for the HashTable and the
lookup time of objects in the cache.

All Mononoke applications are configured to use a formula based upon the size of
the cache and the average size of each object in the cache. The average object
size is assumed to be 200 bytes, however this isn't true for the LFS server.

Update the cachelib helper so that an average object size can be passed.

Further, add a commandline argument (`--buckets-power`) that can be used to
explictly state the number of buckets in cachelib's HashTable.

Reviewed By: krallin

Differential Revision: D19662509

fbshipit-source-id: a27ef37c801af3f835d38f71f88714c07bb27af6