mononoke: bookmarks validator for megarepo
Summary:
We are going to enable megarepo setup soon, and we'd like to verify that
bookmarks are the same.

We do it by comparing the bookmarks between small and large repo, and if they are different then we check the history of this bookmark in the large repo. If it recently pointed to the equivalent commit from the small repo then we assume that validation was successful.

Reviewed By: krallin

Differential Revision: D19623843

fbshipit-source-id: 99e399e43341fe4bca563560b2523994dc19fc53