mononoke: remove filenodes generation from CreateChangeset
Summary:
We have two ways to generate hg changesets in Mononoke: generate new hg
changesets from a bonsai changeset (see
BlobRepo::get_hg_from_bonsai_changeset() function) or accept a commit
generated by Mercurial either via blobimport or through the wireproto unbundle
command(CreateChangeset).

The main goal of the stack is to remove filenode generation from both of those
places, and this diff cleans up the CreateChangeset. By doing it we are able to get rid
of an ugly hack of specifying whether a commit is draft or not.

Reviewed By: krallin

Differential Revision: D19516956

fbshipit-source-id: 77921612a919a52f6ce1ddfbf55944ca300db46e