mononoke: walker: fix race when loading phases
Summary:
The is_public check calls phases.get_public_derive, which in turn calls repo.get_bonsai_heads_maybe_stale. The later is a race as the heads may be out of date vs the changesets being walked if the walker is very near a recent commit.

This change loads the heads information once, so that it can be used in calls to get_public_derive throughout the walk

Also removes the SqlPhases dependency, walker now works using dyn Phases

Reviewed By: krallin

Differential Revision: D19601580

fbshipit-source-id: 31779c4d858cb4fa3ca98dc715d624a6b2ad8cd6