mononoke: disable filenode generation during hg changeset generation
Summary:
Now that we have FilenodesOnlyPublic derived data we no longer need to generate
filenods when we generate hg changesets.

This is important, because we want to generate hg changesets for both public
and draft commits, but we want to generate filenodes only for public commits.
By separating filenodes generation from hg changeset generation we can finally
achieve that.

This diff is the final step - it removes IncompleteFilenodes struct completely.

A few caveats:
1) This diff also removes regenerate_filenodes binary - backfill_derived_data can
be used instead.
2) In D18748102 we added a hack that force-backfilled filenodes when a merge
commit is pushed. This hack is no longer necessary and it was removed, but we
still want to protect against pushing too many commits with no filenodes
generated for them. So instead of force backfilling the filenodes the code now
just checks how commits are missing filenodes.

Reviewed By: aslpavel

Differential Revision: D19683369

fbshipit-source-id: adf613defa50400e55a95dbe14053d954b88016b