mononoke: add count_underived() function to derived data
Summary:
This function can be used to count how many ancestors of a commit do not have
data derived. It will be used in the next diff

Note that count_underived() function traverses the commit graph, so it can be
quite expensive. That's why `limit` parameter is required to bound the walk.

Reviewed By: farnz

Differential Revision: D19683368

fbshipit-source-id: 2ca07db19fc10b5ea86107b084a28072ec4703fa