mononoke: fix test
Summary:
The idea of the test was that this should be a stack of empty commits, and deriving
the top one also derives the bottom commit. The test happened to pass because
we still generate filenodes during hg changeset generation, however it started
to break in the next diff. So this diff fixes the test.

While at it also fix the comment

Reviewed By: farnz

Differential Revision: D19690293

fbshipit-source-id: d46f5686c8468b5a180104c9e5313e51eafc3092