common/rust/manifold: convert to async / await
Summary:
The underlying Thrift client code is already async / await, but the layer on
top of it is not. This isn't ideal, since it results in a lot of undesirable
clones happening all over the place (e.g. every key is cloned multiple times).

This fixes that by updating the wrapper code to async / await.

Reviewed By: farnz

Differential Revision: D19695708

fbshipit-source-id: b4c1b6fd6bf4d998ca6187c4cdde67294329d181